added static windowMap for hwnd<>Window* lookup in windows implementation, removed all compiler warnings (except thirdparty stuff)

+++ projects/lostengine/lost/bitmap/Bitmap.cpp
@@ -397,8 +397,8 @@ void Bitmap::drawRectOutline(const lost::math::Rect& inRect, const common::Color
 
 void Bitmap::disc(float x, float y, float r)
 {
-  float w = width;
-  float h = height;
+  float w = (float)width;
+  float h = (float)height;
   common::Color white(1,1,1,1);
   common::Color black(0,0,0,0);
   for(float cx=0; cx<w; cx+=1.0f)
@@ -408,12 +408,12 @@ void Bitmap::disc(float x, float y, float r)
       float dx = cx-x;
       float dy = cy-y;
       float cr = sqrtf(dx*dx+dy*dy);
-      pixel(cx,cy,(cr < r) ? white : black);
+      pixel((uint32_t)cx, (uint32_t)cy, (cr < r) ? white : black);
       float dr = cr-r;
       if(dr >= 0.0f && (dr < 1.0f))
       {
         float fract = 1.0f-(dr - floor(dr));
-        pixel(cx,cy,common::Color(1,1,1,fract));
+        pixel((uint32_t)cx, (uint32_t)cy, common::Color(1,1,1,fract));
       }
     }
   }
